pyproject.toml
[project]
name = "blive-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql",
    "aiofiles",
    "aiohttp",
    "python-dotenv",
    "loguru",
    "cryptography",
    "brotli",
    "websockets",
    "pydantic>=2.11.7",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
